test(openai): consolidate realtime voice fixtures
