+++ extensions/openai/realtime-voice-bridge-connection.test.ts
@@ -4,146 +4,31 @@ import type { RealtimeVoiceBridge } from "openclaw/plugin-sdk/realtime-voice";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { buildOpenAIRealtimeVoiceProvider } from "./realtime-voice-provider.js";
 
-const INTERNAL_REALTIME_VOICE_PROVIDER = Symbol.for("openclaw.internal.realtime-voice-provider.v1");
-
-function readInternalRealtimeVoiceProviderApi(provider: object) {
-  return Reflect.get(provider, INTERNAL_REALTIME_VOICE_PROVIDER) as {
-    isBrowserSessionConfigured: (ctx: {
-      cfg?: object;
-      providerConfig: Record<string, unknown>;
-      agentId?: string;
-    }) => boolean;
-    isGatewayRelayConfigured: (ctx: {
-      cfg?: object;
-      providerConfig: Record<string, unknown>;
-      agentId?: string;
-    }) => boolean | undefined;
-    resolveBrowserSessionCapabilities: (ctx: {
-      cfg?: object;
-      providerConfig: Record<string, unknown>;
-      model?: string;
-    }) => {
-      handlesAgentConsult?: boolean;
-      supportsToolCalls?: boolean;
-      supportsVideoFrames?: boolean;
-      supportsGatewayControl?: boolean;
-      transports?: string[];
-    };
-    resolveGatewayRelayCapabilities: (ctx: {
-      cfg?: object;
-      providerConfig: Record<string, unknown>;
-      model?: string;
-    }) => {
-      handlesAgentConsult?: boolean;
-      supportsToolCalls?: boolean;
-      transports?: string[];
-    };
-    validateGatewayRelayLaunch: (ctx: {
-      cfg?: object;
-      providerConfig: Record<string, unknown>;
-      model?: string;
-      autoRespondToAudio?: boolean;
-    }) => string | undefined;
-    cancelBrowserSession: (request: Record<string, unknown>, session: object) => Promise<void>;
-  };
-}
-
-const {
-  FakeWebSocket,
-  execFileSyncMock,
-  fetchWithSsrFGuardMock,
-  isProviderAuthProfileConfiguredMock,
-  resolveProviderAuthProfileApiKeyMock,
-} = vi.hoisted(() => {
-  type Listener = (...args: unknown[]) => void;
-
-  class MockWebSocket {
-    static readonly OPEN = 1;
-    static readonly CLOSED = 3;
-    static instances: MockWebSocket[] = [];
-
-    readonly listeners = new Map<string, Listener[]>();
-    readyState = 0;
-    sent: string[] = [];
-    closed = false;
-    terminated = false;
-    deferClose = false;
-    deferredClose: (() => void) | undefined;
-    args: unknown[];
-
-    constructor(...args: unknown[]) {
-      this.args = args;
-      MockWebSocket.instances.push(this);
-    }
-
-    on(event: string, listener: Listener): this {
-      const listeners = this.listeners.get(event) ?? [];
-      listeners.push(listener);
-      this.listeners.set(event, listeners);
-      return this;
-    }
-
-    emit(event: string, ...args: unknown[]): void {
-      for (const listener of this.listeners.get(event) ?? []) {
-        listener(...args);
-      }
-    }
-
-    send(payload: string): void {
-      this.sent.push(payload);
-    }
-
-    close(code?: number, reason?: string): void {
-      this.closed = true;
-      this.readyState = MockWebSocket.CLOSED;
-      const emitClose = () => this.emit("close", code ?? 1000, Buffer.from(reason ?? ""));
-      if (this.deferClose) {
-        this.deferredClose = emitClose;
-        return;
-      }
-      emitClose();
-    }
-
-    terminate(): void {
-      this.terminated = true;
-      this.close(1006, "terminated");
-    }
-
-    emitDeferredClose(): void {
-      const emitClose = this.deferredClose;
-      this.deferredClose = undefined;
-      emitClose?.();
-    }
-  }
-
-  return {
-    FakeWebSocket: MockWebSocket,
-    execFileSyncMock: vi.fn(),
-    fetchWithSsrFGuardMock: vi.fn(),
-    isProviderAuthProfileConfiguredMock: vi.fn(),
-    resolveProviderAuthProfileApiKeyMock: vi.fn(),
-  };
+const mocks = await vi.hoisted(async () => {
+  const { createOpenAIRealtimeMockState } = await import("./realtime-voice-test-support.js");
+  return createOpenAIRealtimeMockState();
 });
+const { FakeWebSocket, fetchWithSsrFGuardMock } = mocks;
 
 vi.mock("node:child_process", async (importOriginal) => {
   const actual = await importOriginal<typeof import("node:child_process")>();
   return {
     ...actual,
-    execFileSync: execFileSyncMock,
+    execFileSync: mocks.execFileSyncMock,
   };
 });
 
 vi.mock("ws", () => ({
-  default: FakeWebSocket,
+  default: mocks.FakeWebSocket,
 }));
 
 vi.mock("openclaw/plugin-sdk/ssrf-runtime", () => ({
-  fetchWithSsrFGuard: fetchWithSsrFGuardMock,
+  fetchWithSsrFGuard: mocks.fetchWithSsrFGuardMock,
 }));
 
 vi.mock("openclaw/plugin-sdk/provider-auth", () => ({
-  isProviderAuthProfileConfigured: isProviderAuthProfileConfiguredMock,
-  resolveProviderAuthProfileApiKey: resolveProviderAuthProfileApiKeyMock,
+  isProviderAuthProfileConfigured: mocks.isProviderAuthProfileConfiguredMock,
+  resolveProviderAuthProfileApiKey: mocks.resolveProviderAuthProfileApiKeyMock,
 }));
 import { createOpenAIRealtimeTestSupport } from "./realtime-voice-test-support.js";
 
@@ -164,23 +49,19 @@ const {
   createRealtimeTool,
   createUnreadableToolName,
   createMalformedToolName,
-} = createOpenAIRealtimeTestSupport({ FakeWebSocket, fetchWithSsrFGuardMock });
+  resetTestState,
+  restoreTestEnvironment,
+  readInternalRealtimeVoiceProviderApi,
+  createQuicksilverBrowserBrokerFixture,
+} = createOpenAIRealtimeTestSupport({ ...mocks, buildOpenAIRealtimeVoiceProvider });
 
 describe("OpenAI realtime voice bridge connection", () => {
   beforeEach(() => {
-    FakeWebSocket.instances = [];
-    vi.stubEnv("OPENAI_API_KEY", "");
-    execFileSyncMock.mockReset();
-    fetchWithSsrFGuardMock.mockReset();
-    isProviderAuthProfileConfiguredMock.mockReset();
-    isProviderAuthProfileConfiguredMock.mockReturnValue(false);
-    resolveProviderAuthProfileApiKeyMock.mockReset();
-    resolveProviderAuthProfileApiKeyMock.mockResolvedValue(undefined);
+    resetTestState();
   });
 
   afterEach(() => {
-    vi.useRealTimers();
-    vi.unstubAllEnvs();
+    restoreTestEnvironment();
   });
 
   it("adds OpenClaw attribution headers to native realtime websocket requests", () => {
@@ -209,21 +90,11 @@ describe("OpenAI realtime voice bridge connection", () => {
   });
 
   it("sends one shared GA policy and waits for session.updated on an attached sideband", async () => {
-    const createBrowserSession = vi.fn(
-      async (_request: unknown, _auth: unknown) =>
-        ({
-          provider: "openai",
-          transport: "webrtc" as const,
-          clientSecret: "gateway-token",
-          offerUrl: "/plugins/openai/realtime/calls",
-        }) as const,
-    );
+    const { broker, createBrowserSession } = createQuicksilverBrowserBrokerFixture({
+      session: { clientSecret: "gateway-token" },
+    });
     const provider = buildOpenAIRealtimeVoiceProvider({
-      quicksilverBrowserSessionBroker: {
-        capabilities: { handlesAgentConsult: true as const },
-        createBrowserSession,
-        cancelBrowserSession: vi.fn(async () => undefined),
-      },
+      quicksilverBrowserSessionBroker: broker,
     });
     const bindBridge = vi.fn();
     const onEvent = vi.fn();
@@ -620,15 +491,10 @@ describe("OpenAI realtime voice bridge connection", () => {
     const { connecting, socket } = beginBridgeConnection(bridge);
 
     openSocket(socket);
-    socket.emit(
-      "message",
-      Buffer.from(
-        JSON.stringify({
-          type: "error",
-          error: { message: "invalid realtime session" },
-        }),
-      ),
-    );
+    emitServerEvent(socket, {
+      type: "error",
+      error: { message: "invalid realtime session" },
+    });
 
     await expect(connecting).rejects.toThrow("invalid realtime session");
     expect(bridge.isConnected()).toBe(false);
